add tests for in expressions
The standard defines IN expressions as equivalent to =ANY expressions, which are
themselves defined under <quantified comparison predicate>. The quantified
comparisons use ternary logic. The handling of ternary logic is complicated,
confusing, and error prone.  So we add tests for compliance to see whether
systems handle IN expressions correctly.

We separate the tests into those with a single attribute and those with row expressions
as there are many systems that only support IN expressions with a single attribute.